refactor(navigator): Stack
